requirements.txt
fastapi
uvicorn>=0.15.0,<0.16.0

python-dotenv

sqlalchemy
psycopg2-binary

pydantic
tenacity

python-jose[cryptography]
passlib[bcrypt]
python-multipart

redis

pytest
pytest-cov
pytest-mock
requests

gunicorn

pytz